requirements.txt
tensorflow
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
